Split Junit module between base utils and the integration test runtime
