+++ core/common/lib/keys-lib/build.gradle.kts
@@ -25,7 +25,7 @@ dependencies {
     implementation(libs.nimbus.jwt)
     implementation(libs.tink)
 
-    testImplementation(project(":core:common:junit"));
+    testImplementation(project(":tests:junit-base"));
 
 }
