+++ extensions/common/iam/decentralized-identity/identity-did-web/src/test/java/org/eclipse/edc/iam/did/web/resolution/WebDidResolverTest.java
@@ -28,7 +28,7 @@ import java.nio.charset.StandardCharsets;
 
 import static okhttp3.Protocol.HTTP_1_1;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.edc.junit.testfixtures.TestUtils.testHttpClient;
+import static org.eclipse.http.client.testfixtures.HttpTestUtils.testHttpClient;
 import static org.mockito.Mockito.mock;
 
 class WebDidResolverTest {
