+++ extensions/common/transaction/transaction-local/build.gradle.kts
@@ -22,7 +22,8 @@ dependencies {
     api(project(":spi:common:transaction-spi"))
     implementation(project(":spi:common:transaction-datasource-spi"))
 
-    testImplementation(project(":core:common:junit"))
+    testImplementation(project(":tests:junit-base"));
+
     testImplementation(testFixtures(project(":extensions:common:sql:sql-core")))
 }
