+++ extensions/common/validator/validator-data-address-kafka/build.gradle.kts
@@ -22,5 +22,6 @@ dependencies {
     api(project(":spi:common:validator-spi"))
     api(project(":spi:common:data-address:data-address-kafka-spi"))
 
-    testImplementation(project(":core:common:junit"))
+    testImplementation(project(":tests:junit-base"));
+
 }
