+++ extensions/common/vault/vault-hashicorp/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
     implementation(project(":core:common:util"))
 
     testImplementation(project(":core:common:connector-core"))
+    testImplementation(testFixtures(project(":core:common:lib:http-lib")))
     testImplementation(project(":core:common:junit"))
     testImplementation(libs.jakartaJson)
     testImplementation(libs.testcontainers.junit)
