+++ extensions/control-plane/callback/callback-http-dispatcher/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
 
     testImplementation(libs.mockserver.netty)
     testImplementation(libs.mockserver.client)
+    testImplementation(testFixtures(project(":core:common:lib:http-lib")))
     testImplementation(project(":core:common:junit"))
 
 }
