+++ extensions/control-plane/transfer/transfer-pull-http-dynamic-receiver/build.gradle.kts
@@ -28,7 +28,8 @@ dependencies {
     testImplementation(project(":spi:common:data-address:data-address-http-data-spi"))
     testImplementation(libs.mockserver.netty)
     testImplementation(libs.mockserver.client)
-    testImplementation(project(":core:common:junit"))
+
+    testImplementation(testFixtures(project(":core:common:lib:http-lib")))
 
 
 }
