+++ extensions/control-plane/transfer/transfer-pull-http-dynamic-receiver/src/test/java/org/eclipse/edc/connector/receiver/http/dynamic/HttpDynamicEndpointDataReferenceReceiverTest.java
@@ -40,8 +40,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.eclipse.edc.connector.receiver.http.dynamic.HttpDynamicEndpointDataReferenceReceiver.HTTP_RECEIVER_ENDPOINT;
 import static org.eclipse.edc.connector.receiver.http.dynamic.TestFunctions.createTransferProcess;
 import static org.eclipse.edc.connector.receiver.http.dynamic.TestFunctions.transferProperties;
-import static org.eclipse.edc.junit.testfixtures.TestUtils.testOkHttpClient;
 import static org.eclipse.edc.util.io.Ports.getFreePort;
+import static org.eclipse.http.client.testfixtures.HttpTestUtils.testOkHttpClient;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.atMostOnce;
