+++ extensions/control-plane/transfer/transfer-pull-http-receiver/src/test/java/org/eclipse/edc/connector/receiver/http/HttpEndpointDataReferenceReceiverTest.java
@@ -34,8 +34,8 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.edc.junit.testfixtures.TestUtils.testHttpClient;
 import static org.eclipse.edc.util.io.Ports.getFreePort;
+import static org.eclipse.http.client.testfixtures.HttpTestUtils.testHttpClient;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockserver.integration.ClientAndServer.startClientAndServer;
