+++ extensions/data-plane/data-plane-signaling/data-plane-signaling-client/build.gradle.kts
@@ -34,6 +34,8 @@ dependencies {
     testImplementation(libs.restAssured)
     testImplementation(libs.mockserver.netty)
     testImplementation(libs.mockserver.client)
+
+    testImplementation(testFixtures(project(":core:common:lib:http-lib")))
 }
